Modify uint64_t to unsigned long long

+++ vic/drivers/shared_all/src/vic_log.c
@@ -52,7 +52,7 @@ get_current_datetime(char *cdt)
 {
     char         ymd[MAXSTRING];
     struct tm   *timeinfo;
-    uint64_t seconds_since_midnight;
+    unsigned long long seconds_since_midnight;
     time_t       curr_date_time;
 
     curr_date_time = time(NULL);
@@ -62,7 +62,7 @@ get_current_datetime(char *cdt)
 
     timeinfo = localtime(&curr_date_time);
 
-    seconds_since_midnight = (uint64_t) curr_date_time % CONST_CDAY;
+    seconds_since_midnight = (unsigned long long) curr_date_time % CONST_CDAY;
 
     if (strftime(ymd, MAXSTRING - 1, "%Y%m%d", timeinfo) == 0) {
         log_err("Something went wrong converting the current time info to ymd");
